src/test/java/ru/portfolio/portfolio/parser/psb/CouponAndAmortizationTableTest.java
package ru.portfolio.portfolio.parser.psb;

import org.testng.annotations.DataProvider;
import org.testng.annotations.Test;

import java.io.IOException;
import java.util.List;

import static org.testng.Assert.assertEquals;

public class CouponAndAmortizationTableTest {
    PsbBrokerReport report;

    CouponAndAmortizationTableTest() throws IOException {
        this.report = new PsbBrokerReport("E:\\1.xlsx");
    }

    @DataProvider(name = "isin")
    Object[][] getData() {
        return new Object[][] {{"RU000A0ZYAQ7", "RU000A0JV3M2" }};
    }

    @Test(dataProvider = "isin")
    void testIsin(String firstIsin, String lastIsin) {
        List<CouponAndAmortizationTable.CouponAndAmortizationTableRow> data = new CouponAndAmortizationTable(this.report).getData();
        assertEquals(data.get(0).getIsin(), firstIsin);
        assertEquals(data.get(data.size() - 1).getIsin(), lastIsin);
    }
}